Refactor extract LoginDialog to its own file
