+++ securedrop_client/gui/widgets.py
@@ -18,7 +18,6 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 """
 import html
 import logging
-import sys
 from gettext import gettext as _
 from gettext import ngettext
 from typing import Dict, List, Optional, Union  # noqa: F401
@@ -29,7 +28,6 @@ import sqlalchemy.orm.exc
 from PyQt5.QtCore import QEvent, QObject, QSize, Qt, QTimer, pyqtBoundSignal, pyqtSignal, pyqtSlot
 from PyQt5.QtGui import (
     QBrush,
-    QCloseEvent,
     QColor,
     QCursor,
     QFocusEvent,
@@ -49,7 +47,6 @@ from PyQt5.QtWidgets import (
     QDialog,
     QDialogButtonBox,
     QGraphicsDropShadowEffect,
-    QGraphicsOpacityEffect,
     QGridLayout,
     QHBoxLayout,
     QLabel,
@@ -68,7 +65,6 @@ from PyQt5.QtWidgets import (
     QWidget,
 )
 
-from securedrop_client import __version__ as sd_version
 from securedrop_client.db import (
     DraftReply,
     File,
@@ -1695,202 +1691,6 @@ class PasswordEdit(QLineEdit):
             self.setEchoMode(QLineEdit.Password)
             self.password_shown = False
             self.togglepasswordAction.setIcon(self.hiddenIcon)
-
-
-class LoginDialog(QDialog):
-    """
-    A dialog to display the login form.
-    """
-
-    MIN_PASSWORD_LEN = 14  # Journalist.MIN_PASSWORD_LEN on server
-    MAX_PASSWORD_LEN = 128  # Journalist.MAX_PASSWORD_LEN on server
-    MIN_JOURNALIST_USERNAME = 3  # Journalist.MIN_USERNAME_LEN on server
-
-    def __init__(self, parent) -> None:  # type: ignore [no-untyped-def]
-        self.parent = parent
-        super().__init__(self.parent)
-
-        # Set modal
-        self.setModal(True)
-
-        # Set layout
-        layout = QVBoxLayout(self)
-        self.setLayout(layout)
-
-        # Set margins and spacing
-        layout.setContentsMargins(0, 274, 0, 20)
-        layout.setSpacing(0)
-
-        # Set background
-        self.setAutoFillBackground(True)
-        palette = QPalette()
-        palette.setBrush(QPalette.Background, QBrush(load_image("login_bg.svg")))
-        self.setPalette(palette)
-        self.setFixedSize(QSize(596, 671))  # Set to size provided in the login_bg.svg file
-        self.setSizePolicy(QSizePolicy.Fixed, QSizePolicy.Fixed)
-
-        # Create error bar
-        self.error_bar = LoginErrorBar()
-
-        # Create form widget
-        form = QWidget()
-
-        form.setObjectName("LoginDialog_form")
-
-        form_layout = QVBoxLayout()
-        form.setLayout(form_layout)
-
-        form_layout.setContentsMargins(80, 0, 80, 0)
-        form_layout.setSpacing(8)
-
-        self.username_label = QLabel(_("Username"))
-        self.username_field = QLineEdit()
-
-        self.password_label = QLabel(_("Passphrase"))
-        self.password_field = PasswordEdit(self)
-
-        self.tfa_label = QLabel(_("Two-Factor Code"))
-        self.tfa_field = QLineEdit()
-
-        self.opacity_effect = QGraphicsOpacityEffect()
-
-        buttons = QWidget()
-        buttons_layout = QHBoxLayout()
-        buttons.setLayout(buttons_layout)
-        buttons_layout.setContentsMargins(0, 20, 0, 0)
-        self.submit = SignInButton()
-        self.submit.clicked.connect(self.validate)
-        self.offline_mode = LoginOfflineLink()
-        buttons_layout.addWidget(self.offline_mode)
-        buttons_layout.addStretch()
-        buttons_layout.addWidget(self.submit)
-
-        form_layout.addWidget(self.username_label)
-        form_layout.addWidget(self.username_field)
-        form_layout.addWidget(QWidget(self))
-        form_layout.addWidget(self.password_label)
-        form_layout.addWidget(self.password_field)
-        form_layout.addWidget(QWidget(self))
-        form_layout.addWidget(self.tfa_label)
-        form_layout.addWidget(self.tfa_field)
-        form_layout.addWidget(buttons)
-
-        # Create widget to display application name and version
-        application_version = QWidget()
-        application_version_layout = QHBoxLayout()
-        application_version.setLayout(application_version_layout)
-        application_version_label = QLabel(_("SecureDrop Client v{}").format(sd_version))
-        application_version_label.setAlignment(Qt.AlignHCenter)
-        application_version_label.setObjectName("LoginDialog_app_version_label")
-        application_version_layout.addWidget(application_version_label)
-
-        # Add widgets
-        layout.addWidget(self.error_bar)
-        layout.addStretch()
-        layout.addWidget(form)
-        layout.addStretch()
-        layout.addWidget(application_version)
-
-    def closeEvent(self, event: QCloseEvent) -> None:
-        """
-        Only exit the application when the main window is not visible.
-        """
-        if not self.parent.isVisible():
-            sys.exit(0)
-
-    def keyPressEvent(self, event: QKeyEvent) -> None:
-        """
-        Cutomize keyboard behavior in the login dialog.
-
-        - [Esc] should not close the dialog
-        - [Enter] or [Return] should attempt to submit the form
-        """
-        if event.key() == Qt.Key_Escape:
-            event.ignore()
-
-        if event.key() == Qt.Key_Enter or event.key() == Qt.Key_Return:
-            self.validate()
-
-    def setup(self, controller: Controller) -> None:
-        self.controller = controller
-        self.offline_mode.clicked.connect(self.controller.login_offline_mode)
-
-    def reset(self) -> None:
-        """
-        Resets the login form to the default state.
-        """
-        self.username_field.setText("")
-        self.username_field.setFocus()
-        self.password_field.setText("")
-        self.tfa_field.setText("")
-        self.setDisabled(False)
-        self.error_bar.clear_message()
-
-    def error(self, message: str) -> None:
-        """
-        Ensures the passed in message is displayed as an error message.
-        """
-        self.setDisabled(False)
-        self.submit.setText(_("SIGN IN"))
-        self.error_bar.set_message(message)
-
-        self.opacity_effect.setOpacity(1)
-        self.offline_mode.setGraphicsEffect(self.opacity_effect)
-
-    def validate(self) -> None:
-        """
-        Validate the user input -- we expect values for:
-
-        * username (free text)
-        * password (free text)
-        * TFA token (numerals)
-        """
-        self.setDisabled(True)
-        username = self.username_field.text()
-        password = self.password_field.text()
-        tfa_token = self.tfa_field.text().replace(" ", "")
-        if username and password and tfa_token:
-            # Validate username
-            if len(username) < self.MIN_JOURNALIST_USERNAME:
-                self.setDisabled(False)
-                self.error(
-                    _("That username won't work.\n" "It should be at least 3 characters long.")
-                )
-                return
-
-            # Validate password
-            if len(password) < self.MIN_PASSWORD_LEN or len(password) > self.MAX_PASSWORD_LEN:
-                self.setDisabled(False)
-                self.error(
-                    _(
-                        "That passphrase won't work.\n"
-                        "It should be between 14 and 128 characters long."
-                    )
-                )
-                return
-
-            # Validate 2FA token
-            try:
-                int(tfa_token)
-            except ValueError:
-                self.setDisabled(False)
-                self.error(
-                    _("That two-factor code won't work.\n" "It should only contain numerals.")
-                )
-                return
-            self.submit.setText(_("SIGNING IN"))
-
-            # Changing the opacity of the link to .4 alpha of its' 100% value when authenticated
-            self.opacity_effect.setOpacity(0.4)
-            self.offline_mode.setGraphicsEffect(self.opacity_effect)
-
-            # If authentication is successful, clear error messages displayed priorly
-            self.error_bar.clear_message()
-
-            self.controller.login(username, password, tfa_token)
-        else:
-            self.setDisabled(False)
-            self.error(_("Please enter a username, passphrase and " "two-factor code."))
 
 
 class SenderIcon(QWidget):
